move ns tree modal notifs
