+++ src/frontend/src/common/components/library/data_subview/components/CreateNSTreeModal/index.tsx
@@ -1,10 +1,15 @@
 import RadioGroup from "@material-ui/core/RadioGroup";
 import CloseIcon from "@material-ui/icons/Close";
 import { uniq } from "lodash";
+import NextLink from "next/link";
 import React, { SyntheticEvent, useEffect, useState } from "react";
 import { NewTabLink } from "src/common/components/library/NewTabLink";
 import { useCreateTree } from "src/common/queries/trees";
+import { ROUTES } from "src/common/routes";
+import { B } from "src/common/styles/support/style";
 import { pluralize } from "src/common/utils/strUtils";
+import Notification from "src/components/Notification";
+import { ContactUsLink } from "../ContactUsLink";
 import { Header, StyledIconButton } from "../DownloadModal/style";
 import { FailedSampleAlert } from "../FailedSampleAlert";
 import { CreateTreeButton } from "./components/CreateTreeButton";
@@ -19,6 +24,7 @@ import {
   CreateTreeInfo,
   FieldTitle,
   Separator,
+  StyledButton,
   StyledDialog,
   StyledDialogContent,
   StyledDialogTitle,
@@ -52,14 +58,18 @@ export const CreateNSTreeModal = ({
   failedSampleIds,
   open,
   onClose,
-  handleCreateTreeFailed,
-  handleSetCreateTreeStarted,
 }: Props): JSX.Element => {
   const [treeName, setTreeName] = useState<string>("");
   const [isInputInEditMode, setIsInputInEditMode] = useState<boolean>(false);
   const [shouldReset, setShouldReset] = useState<boolean>(false);
   const [treeType, setTreeType] = useState<TreeType>(TreeTypes.Targeted);
   const [missingInputSamples, setMissingInputSamples] = useState<string[]>([]);
+  const [shouldShowErrorNotification, setShouldShowErrorNotification] =
+    useState<boolean>(false);
+  const [
+    shouldShowTreeCreatedNotification,
+    setShouldShowTreeCreatedNotification,
+  ] = useState<boolean>(false);
   const [validatedInputSamples, setValidatedInputSamples] = useState<string[]>(
     []
   );
@@ -98,11 +108,11 @@ export const CreateNSTreeModal = ({
 
   const mutation = useCreateTree({
     onError: () => {
-      handleCreateTreeFailed();
+      setShouldShowErrorNotification(true);
       handleClose();
     },
     onSuccess: () => {
-      handleSetCreateTreeStarted();
+      setShouldShowTreeCreatedNotification(true);
       handleClose();
     },
   });
@@ -135,87 +145,124 @@ export const CreateNSTreeModal = ({
     });
   };
 
+  const handleDismissErrorNotification = () => {
+    setShouldShowErrorNotification(false);
+  };
+
   return (
-    <StyledDialog
-      disableBackdropClick
-      disableEscapeKeyDown
-      open={open}
-      onClose={handleClose}
-      fullWidth={true}
-      maxWidth={"sm"}
-    >
-      <StyledDialogTitle>
-        <StyledIconButton onClick={handleClose}>
-          <CloseIcon />
-        </StyledIconButton>
-        <Header>Create New Phylogenetic Tree</Header>
-        <Title>
-          {allSamplesRequestedTableAndInput.length}{" "}
-          {pluralize("Sample", allValidSamplesForTreeCreation.length)} Total
-        </Title>
-      </StyledDialogTitle>
-      <StyledDialogContent data-test-id="modal-content">
-        <TreeNameInput setTreeName={setTreeName} treeName={treeName} />
-        <TreeTypeSection>
-          <TreeNameInfoWrapper>
-            <FieldTitle>Tree Type: </FieldTitle>
-            <StyledTooltip
-              arrow
-              leaveDelay={1000}
-              title={TREE_TYPE_TOOLTIP_TEXT}
-              placement="top"
+    <>
+      <Notification
+        buttonOnClick={handleDismissErrorNotification}
+        buttonText="DISMISS"
+        dismissDirection="right"
+        dismissed={!shouldShowErrorNotification}
+        intent="error"
+      >
+        <B>Something went wrong and we were unable to start your tree build</B>{" "}
+        <ContactUsLink />
+      </Notification>
+      <Notification
+        autoDismiss={12000}
+        dismissDirection="right"
+        dismissed={!shouldShowTreeCreatedNotification}
+        intent="info"
+      >
+        <span>
+          Your tree is being created. It may take up to 12 hours to process. To
+          check your tree’s status, visit the Phylogenetic Tree tab.
+        </span>
+        <NextLink href={ROUTES.PHYLO_TREES} passHref>
+          <a href="passRef">
+            <StyledButton
+              color="primary"
+              onClick={() => setShouldShowTreeCreatedNotification(false)}
             >
-              <StyledInfoOutlinedIcon />
-            </StyledTooltip>
-          </TreeNameInfoWrapper>
-          <RadioGroup
-            value={treeType}
-            onChange={(e) => setTreeType(e.target.value as TreeType)}
-          >
-            <StyledFormControlLabel
-              value={TreeTypes.Targeted}
-              checked={treeType === TreeTypes.Targeted}
-              control={<StyledRadio />}
-              label={
-                <RadioLabelTargeted
-                  selected={treeType === TreeTypes.Targeted}
-                />
-              }
-            />
-            <StyledFormControlLabel
-              value={TreeTypes.NonContextualized}
-              checked={treeType === TreeTypes.NonContextualized}
-              control={<StyledRadio />}
-              label={
-                <RadioLabelNonContextualized
-                  selected={treeType === TreeTypes.NonContextualized}
-                />
-              }
-            />
-          </RadioGroup>
-        </TreeTypeSection>
-        <Separator marginSize="l" />
-        <SampleIdInput
-          handleInputModeChange={handleInputModeChange}
-          handleInputValidation={handleInputValidation}
-          shouldReset={shouldReset}
-        />
-        <Separator marginSize="xl" />
-        <MissingSampleAlert missingSamples={missingInputSamples} />
-        <FailedSampleAlert numFailedSamples={failedSampleIds?.length} />
-      </StyledDialogContent>
-      <StyledFooter>
-        <CreateTreeButton
-          hasValidName={hasValidName}
-          hasSamples={allValidSamplesForTreeCreation.length > 0}
-          isInEditMode={isInputInEditMode}
-          isValidTreeType={Object.values(TreeTypes).includes(treeType)}
-          onClick={handleSubmit}
-        />
-        <CreateTreeInfo>
-          Creating a new tree can take up to 12 hours.
-        </CreateTreeInfo>
-      </StyledFooter>
-    </StyledDialog>
+              VIEW MY TREES
+            </StyledButton>
+          </a>
+        </NextLink>
+      </Notification>
+      <StyledDialog
+        disableBackdropClick
+        disableEscapeKeyDown
+        open={open}
+        onClose={handleClose}
+        fullWidth={true}
+        maxWidth={"sm"}
+      >
+        <StyledDialogTitle>
+          <StyledIconButton onClick={handleClose}>
+            <CloseIcon />
+          </StyledIconButton>
+          <Header>Create New Phylogenetic Tree</Header>
+          <Title>
+            {allSamplesRequestedTableAndInput.length}{" "}
+            {pluralize("Sample", allValidSamplesForTreeCreation.length)} Total
+          </Title>
+        </StyledDialogTitle>
+        <StyledDialogContent data-test-id="modal-content">
+          <TreeNameInput setTreeName={setTreeName} treeName={treeName} />
+          <TreeTypeSection>
+            <TreeNameInfoWrapper>
+              <FieldTitle>Tree Type: </FieldTitle>
+              <StyledTooltip
+                arrow
+                leaveDelay={1000}
+                title={TREE_TYPE_TOOLTIP_TEXT}
+                placement="top"
+              >
+                <StyledInfoOutlinedIcon />
+              </StyledTooltip>
+            </TreeNameInfoWrapper>
+            <RadioGroup
+              value={treeType}
+              onChange={(e) => setTreeType(e.target.value as TreeType)}
+            >
+              <StyledFormControlLabel
+                value={TreeTypes.Targeted}
+                checked={treeType === TreeTypes.Targeted}
+                control={<StyledRadio />}
+                label={
+                  <RadioLabelTargeted
+                    selected={treeType === TreeTypes.Targeted}
+                  />
+                }
+              />
+              <StyledFormControlLabel
+                value={TreeTypes.NonContextualized}
+                checked={treeType === TreeTypes.NonContextualized}
+                control={<StyledRadio />}
+                label={
+                  <RadioLabelNonContextualized
+                    selected={treeType === TreeTypes.NonContextualized}
+                  />
+                }
+              />
+            </RadioGroup>
+          </TreeTypeSection>
+          <Separator marginSize="l" />
+          <SampleIdInput
+            handleInputModeChange={handleInputModeChange}
+            handleInputValidation={handleInputValidation}
+            shouldReset={shouldReset}
+          />
+          <Separator marginSize="xl" />
+          <MissingSampleAlert missingSamples={missingInputSamples} />
+          <FailedSampleAlert numFailedSamples={failedSampleIds?.length} />
+        </StyledDialogContent>
+        <StyledFooter>
+          <CreateTreeButton
+            hasValidName={hasValidName}
+            hasSamples={allValidSamplesForTreeCreation.length > 0}
+            isInEditMode={isInputInEditMode}
+            isValidTreeType={Object.values(TreeTypes).includes(treeType)}
+            onClick={handleSubmit}
+          />
+          <CreateTreeInfo>
+            Creating a new tree can take up to 12 hours.
+          </CreateTreeInfo>
+        </StyledFooter>
+      </StyledDialog>
+    </>
   );
 };
